requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt setup.py
#
aiohttp==3.6.2            # via slackclient
alembic==1.4.2            # via flask-migrate
amqp==2.6.0               # via kombu
apispec[yaml]==3.3.1      # via flask-appbuilder
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp, jsonschema
babel==2.8.0              # via flask-babel
backoff==1.10.0           # via apache-superset (setup.py)
billiard==3.6.3.0         # via celery
bleach==3.1.5             # via apache-superset (setup.py)
brotli==1.0.7             # via flask-compress
cachelib==0.1.1           # via apache-superset (setup.py)
celery==4.4.6             # via apache-superset (setup.py)
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via aiohttp
click==7.1.2              # via apache-superset (setup.py), flask, flask-appbuilder
colorama==0.4.3           # via apache-superset (setup.py), flask-appbuilder
contextlib2==0.6.0.post1  # via apache-superset (setup.py)
croniter==0.3.31          # via apache-superset (setup.py)
cryptography==2.9.2       # via apache-superset (setup.py)
dataclasses==0.6          # via apache-superset (setup.py)
decorator==4.4.2          # via retry
defusedxml==0.6.0         # via python3-openid
dnspython==1.16.0         # via email-validator
email-validator==1.1.0    # via flask-appbuilder
flask-appbuilder==3.0.0   # via apache-superset (setup.py)
flask-babel==1.0.0        # via flask-appbuilder
flask-caching==1.8.0      # via apache-superset (setup.py)
flask-compress==1.5.0     # via apache-superset (setup.py)
flask-jwt-extended==3.24.1  # via flask-appbuilder
flask-login==0.4.1        # via flask-appbuilder
flask-migrate==2.5.3      # via apache-superset (setup.py)
flask-openid==1.2.5       # via flask-appbuilder
flask-sqlalchemy==2.4.1   # via flask-appbuilder, flask-migrate
flask-talisman==0.7.0     # via apache-superset (setup.py)
flask-wtf==0.14.3         # via apache-superset (setup.py), flask-appbuilder
flask==1.1.2              # via apache-superset (setup.py), flask-appbuilder, flask-babel, flask-caching, flask-compress, flask-jwt-extended, flask-login, flask-migrate, flask-openid, flask-sqlalchemy, flask-wtf
future==0.18.2            # via celery
geographiclib==1.50       # via geopy
geopy==1.22.0             # via apache-superset (setup.py)
gunicorn==20.0.4          # via apache-superset (setup.py)
humanize==2.4.0           # via apache-superset (setup.py)
idna-ssl==1.1.0           # via aiohttp
idna==2.9                 # via email-validator, idna-ssl, yarl
importlib-metadata==1.6.0  # via jsonschema, kombu, markdown
isodate==0.6.0            # via apache-superset (setup.py)
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.11.2            # via flask, flask-babel
jsonschema==3.2.0         # via flask-appbuilder
kombu==4.6.11             # via celery
mako==1.1.2               # via alembic
markdown==3.2.2           # via apache-superset (setup.py)
markupsafe==1.1.1         # via jinja2, mako, wtforms
marshmallow-enum==1.5.1   # via flask-appbuilder
marshmallow-sqlalchemy==0.23.0  # via flask-appbuilder
marshmallow==3.6.1        # via flask-appbuilder, marshmallow-enum, marshmallow-sqlalchemy
msgpack==1.0.0            # via apache-superset (setup.py)
multidict==4.7.6          # via aiohttp, yarl
numpy==1.18.4             # via pandas, pyarrow
packaging==20.3           # via bleach
pandas==1.0.3             # via apache-superset (setup.py)
parsedatetime==2.5        # via apache-superset (setup.py)
pathlib2==2.3.5           # via apache-superset (setup.py)
polyline==1.4.0           # via apache-superset (setup.py)
prison==0.1.3             # via flask-appbuilder
py==1.8.1                 # via retry
pyarrow==0.17.0           # via apache-superset (setup.py)
pycparser==2.20           # via cffi
pyjwt==1.7.1              # via flask-appbuilder, flask-jwt-extended
pyparsing==2.4.7          # via packaging
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via alembic, apache-superset (setup.py), croniter, flask-appbuilder, pandas
python-dotenv==0.13.0     # via apache-superset (setup.py)
python-editor==1.0.4      # via alembic
python-geohash==0.8.5     # via apache-superset (setup.py)
python3-openid==3.1.0     # via flask-openid
pytz==2020.1              # via babel, celery, flask-babel, pandas
pyyaml==5.3.1             # via apache-superset (setup.py), apispec
retry==0.9.2              # via apache-superset (setup.py)
selenium==3.141.0         # via apache-superset (setup.py)
simplejson==3.17.0        # via apache-superset (setup.py)
six==1.14.0               # via bleach, cryptography, flask-jwt-extended, flask-talisman, isodate, jsonschema, packaging, pathlib2, polyline, prison, pyrsistent, python-dateutil, sqlalchemy-utils, wtforms-json
slackclient==2.6.2        # via apache-superset (setup.py)
sqlalchemy-utils==0.36.6  # via apache-superset (setup.py), flask-appbuilder
sqlalchemy==1.3.16        # via alembic, apache-superset (setup.py), flask-sqlalchemy, marshmallow-sqlalchemy, sqlalchemy-utils
sqlparse==0.3.0           # via apache-superset (setup.py)
typing-extensions==3.7.4.2  # via aiohttp
urllib3==1.25.9           # via selenium
vine==1.3.0               # via amqp, celery
webencodings==0.5.1       # via bleach
werkzeug==1.0.1           # via flask, flask-jwt-extended
wtforms-json==0.3.3       # via apache-superset (setup.py)
wtforms==2.3.1            # via flask-wtf, wtforms-json
yarl==1.4.2               # via aiohttp
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools

thrift==0.13.0
sasl==0.2.1
thrift_sasl==0.4.2
pyhive==0.6.3

Authlib==0.14.3
redis==3.5.1
Flask-Session==0.3.2
